requirements.txt
cohere
langchain
beautifulsoup4
pinecone-client